move InputText to widgets general
