+++ lib/ui/screens/home/home_screen.dart
@@ -1,3 +1,4 @@
+import 'package:catbreeds/ui/widgets/general/input_text.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -70,37 +71,6 @@ class _Search extends StatelessWidget {
         hintText: 'Search by breed',
         onChange: _homeProvider!.search,
         suffixIconWidget: const Icon(Icons.search),
-      ),
-    );
-  }
-}
-
-class InputText extends StatelessWidget {
-  const InputText({
-    required this.controller,
-    this.onChange,
-    this.hintText,
-    this.suffixIconWidget,
-    super.key,
-  });
-
-  final TextEditingController controller;
-  final String? Function(String?)? onChange;
-  final String? hintText;
-  final Widget? suffixIconWidget;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      child: TextFormField(
-        controller: controller,
-        onChanged: (value) {
-          if (onChange != null) onChange!(value);
-        },
-        decoration: InputDecoration(
-          hintText: hintText,
-          suffixIcon: suffixIconWidget,
-        ),
       ),
     );
   }
